feat: Refactor flag type handling and introduce FlagTypesUtil for better type management
